feat: remove GlobalInfo
