+++ src/stateAggregator/index.ts
@@ -5,10 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-export * from './globalInfoConfig';
-export * from './sfdxDataHandler';
-export * from './types';
-export { GlobalInfoOrgAccessor, OrgAccessor } from './accessors/orgAccessor';
-export { GlobalInfoAliasAccessor, AliasAccessor } from './accessors/aliasAccessor';
-export { GlobaInfoTokenAccessor, TokenAccessor } from './accessors/tokenAccessor';
+export * from './accessors/orgAccessor';
+export * from './accessors/aliasAccessor';
+export * from './accessors/tokenAccessor';
 export * from './stateAggregator';
