+++ test/unit/stateAggregator/accessors/orgAccessorTest.ts
@@ -5,14 +5,12 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { expect } from 'chai';
-import { GlobalInfo, SfOrg, StateAggregator } from '../../../../src/stateAggregator';
+import { StateAggregator } from '../../../../src/stateAggregator';
 import { AuthFields } from '../../../../src/org';
 import { MockTestOrgData, testSetup, uniqid } from '../../../../src/testSetup';
 
 const username = 'espresso@coffee.com';
-const alias = 'MyAlias';
 const org = new MockTestOrgData(uniqid(), { username });
-const token = { token: '123', url: 'https://login.salesforce.com', user: username };
 
 function expectPartialDeepMatch(actual: AuthFields, expected: AuthFields, ignore = ['refreshToken', 'accessToken']) {
   for (const key of ignore) {
@@ -161,89 +159,6 @@ describe('OrgAccessor', () => {
     it('should return a list of auth file names', async () => {
       const stateAggregator = await StateAggregator.getInstance();
       expect(await stateAggregator.orgs.list()).to.deep.equal([`${username}.json`]);
-    });
-  });
-});
-
-describe('GlobalInfoOrgAccessor', () => {
-  const $$ = testSetup();
-
-  beforeEach(async () => {
-    $$.setConfigStubContents('GlobalInfo', {
-      contents: {
-        orgs: { [username]: await org.getConfig() },
-        aliases: { [alias]: username },
-        tokens: { [username]: token },
-      },
-    });
-  });
-
-  describe('getAll', () => {
-    it('should return all the orgs', async () => {
-      const globalInfo = await GlobalInfo.create();
-      const orgs = globalInfo.orgs.getAll();
-      expectPartialDeepMatch(orgs[username], await org.getConfig());
-    });
-  });
-
-  describe('get', () => {
-    it('should return org that corresponds to a username', async () => {
-      const globalInfo = await GlobalInfo.create();
-      const result = globalInfo.orgs.get(username);
-      expectPartialDeepMatch(result, await org.getConfig());
-    });
-  });
-
-  describe('has', () => {
-    it('should return true if org exists', async () => {
-      const globalInfo = await GlobalInfo.create();
-      const result = globalInfo.orgs.has(username);
-      expect(result).to.deep.equal(true);
-    });
-
-    it('should return false if org does not exist', async () => {
-      const globalInfo = await GlobalInfo.create();
-      const result = globalInfo.orgs.has('DOES_NOT_EXIST');
-      expect(result).to.deep.equal(false);
-    });
-  });
-
-  describe('set', () => {
-    it('should set the org', async () => {
-      const globalInfo = await GlobalInfo.create();
-      const newOrg = { username: 'foobar@baz.com', token: '123' } as unknown as SfOrg;
-      globalInfo.orgs.set('foobar@baz.com', newOrg);
-      const result = globalInfo.orgs.has('foobar@baz.com');
-      expect(result).to.be.true;
-    });
-
-    it('should add the username if does not exist on the object', async () => {
-      const globalInfo = await GlobalInfo.create();
-      const newOrg = { token: '123' } as unknown as SfOrg;
-      globalInfo.orgs.set('foobar@baz.com', newOrg);
-      const result = globalInfo.orgs.get('foobar@baz.com');
-      expect(result.username).to.deep.equal('foobar@baz.com');
-    });
-  });
-
-  describe('update', () => {
-    const instanceUrl = 'https://login.salesforce.com';
-    it('should update the org', async () => {
-      const globalInfo = await GlobalInfo.create();
-      const newOrg = { ...org, instanceUrl };
-      globalInfo.orgs.update(username, newOrg);
-      const result = globalInfo.orgs.get(username);
-      expect(result.instanceUrl).to.deep.equal(instanceUrl);
-    });
-
-    it('should add the username if does not exist on the object', async () => {
-      const globalInfo = await GlobalInfo.create();
-      const newOrg = { ...org, instanceUrl };
-      delete newOrg.username;
-
-      globalInfo.orgs.set(username, newOrg);
-      const result = globalInfo.orgs.get(username);
-      expect(result.username).to.deep.equal(username);
     });
   });
 });
