+++ test/unit/stateAggregator/accessors/tokenAccessorTest.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { expect } from 'chai';
-import { GlobalInfo, StateAggregator } from '../../../../src/stateAggregator';
+import { StateAggregator } from '../../../../src/stateAggregator';
 import { MockTestOrgData, testSetup, uniqid } from '../../../../src/testSetup';
 
 const username = 'espresso@coffee.com';
@@ -59,7 +59,12 @@ describe('TokenAccessor', () => {
   describe('set', () => {
     it('should set the token', async () => {
       const stateAggregator = await StateAggregator.getInstance();
-      const newToken = { username: 'foobar@baz.com', token: '123', url: 'https://login.salesforce.com' };
+      const newToken = {
+        username: 'foobar@baz.com',
+        token: '123',
+        url: 'https://login.salesforce.com',
+        timestamp: new Date().toISOString(),
+      };
       stateAggregator.tokens.set('foobar@baz.com', newToken);
       const result = stateAggregator.tokens.has('foobar@baz.com');
       expect(result).to.be.true;
@@ -82,82 +87,6 @@ describe('TokenAccessor', () => {
       const stateAggregator = await StateAggregator.getInstance();
       stateAggregator.tokens.unset(username);
       const result = stateAggregator.tokens.get(username);
-      expect(result).to.be.undefined;
-    });
-  });
-});
-
-describe('GlobalInfoTokenAccessor', () => {
-  const $$ = testSetup();
-
-  beforeEach(async () => {
-    $$.setConfigStubContents('GlobalInfo', {
-      contents: {
-        orgs: { [username]: await org.getConfig() },
-        aliases: {
-          [alias]: username,
-        },
-        tokens: { [username]: token },
-      },
-    });
-  });
-
-  describe('getAll', () => {
-    it('should return all the tokens', async () => {
-      const globalInfo = await GlobalInfo.create();
-      const tokens = globalInfo.tokens.getAll();
-      expect(tokens).to.deep.equal({ [username]: token });
-    });
-  });
-
-  describe('get', () => {
-    it('should return token that corresponds to a username', async () => {
-      const globalInfo = await GlobalInfo.create();
-      const result = globalInfo.tokens.get(username);
-      expect(result).to.deep.equal(token);
-    });
-  });
-
-  describe('has', () => {
-    it('should return true if token exists', async () => {
-      const globalInfo = await GlobalInfo.create();
-      const result = globalInfo.tokens.has(username);
-      expect(result).to.deep.equal(true);
-    });
-
-    it('should return false if token does not exist', async () => {
-      const globalInfo = await GlobalInfo.create();
-      const result = globalInfo.tokens.has('DOES_NOT_EXIST');
-      expect(result).to.deep.equal(false);
-    });
-  });
-
-  describe('set', () => {
-    it('should set the token', async () => {
-      const globalInfo = await GlobalInfo.create();
-      const newToken = { username: 'foobar@baz.com', token: '123', url: 'https://login.salesforce.com' };
-      globalInfo.tokens.set('foobar@baz.com', newToken);
-      const result = globalInfo.tokens.has('foobar@baz.com');
-      expect(result).to.be.true;
-    });
-  });
-
-  describe('update', () => {
-    it('should update the token', async () => {
-      const globalInfo = await GlobalInfo.create();
-      const instanceUrl = 'https://login.salesforce.com';
-      const newToken = { ...token, instanceUrl };
-      globalInfo.tokens.update(username, newToken);
-      const result = globalInfo.tokens.get(username);
-      expect(result.instanceUrl).to.deep.equal(instanceUrl);
-    });
-  });
-
-  describe('unset', () => {
-    it('should remove the token', async () => {
-      const globalInfo = await GlobalInfo.create();
-      globalInfo.tokens.unset(username);
-      const result = globalInfo.tokens.get(username);
       expect(result).to.be.undefined;
     });
   });
